Move DemoDashboard to dashboard dir
